requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
python-dotenv>=1.0.0
openai>=1.0.0
qdrant-client>=1.7.0
cohere
pydantic-settings
tqdm
httpx
python-multipart
slowapi==0.1.9